Revue complète OnImport

+++ tools/radarrscript/onimport.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
-
+set -eu
 # https://wiki.servarr.com/radarr/custom-scripts
 
 ### ATTENTION les chemins sont en référentiel Docker Radarr
 filmspath="/data/Films"
+radarrpath="/data/Servarr/Films"
 trashpath="/data/Servarr/Trash"
 downloadpath="/data/Servarr/Downloads"
 ticketpath="/data/Servarr/Tickets"
@@ -11,13 +12,22 @@ logfile="/data/Servarr/radarr_event.log"
 
 ## Création d'un ticket pour le rejouer sauf si on vient d'un ticket
 echo ">>>>>> $(date) Begin" >> "$logfile"
-if [[ -z "$ticketfile" ]]; then
+echo "[DEBUG] radarr_movie_title=\"$radarr_movie_title\"" >> "$logfile"
+echo "[DEBUG] radarr_movie_year=\"$radarr_movie_year\"" >> "$logfile"
+echo "[DEBUG] @=\"$@\"" >> "$logfile"
+echo "[DEBUG] ticketfile=\"$ticketfile\"" >> "$logfile"
+# printenv >> "$logfile"
+
+if [[ -z "${ticketfile:-}" ]]; then
     echo "[INFO] From Radarr call" >> "$logfile"
-    
+    if [[ -z "$radarr_movie_title" || -z "$radarr_movie_year" ]]; then
+        echo "[ERROR] Abort : missing movie title or year" >> "$logfile"
+        exit 1
+    fi
     moviename="${radarr_movie_title} (${radarr_movie_year})"
     downloadedfile="$1"
     ticketfile="$ticketpath/$moviename.txt"
-    mkdir -p $ticketpath
+    mkdir -p "$ticketpath"
     echo "moviename=\"$moviename\"" >> "$ticketfile"
     echo "downloadedfile=\"$downloadedfile\"" >> "$ticketfile"
     
@@ -29,65 +39,47 @@ fi
 echo "moviename=\"$moviename\"" >> "$logfile"
 echo "downloadedfile=\"$downloadedfile\"" >> "$logfile"
 
-# printenv >> "$logfile"
 
 ## Vérifier que le download est bien présent
-if [[ ! -e "$downloadedfile" ]]; then
-    echo "[ERROR] Abort download not found \"$downloadedfile\"" >> "$logfile"
-    exit 1
-fi
+[[ -e "$downloadedfile" ]] || { echo "[ERROR] Abort download not found \"$downloadedfile\"" >> "$logfile"; exit 1; }
 
-exit 0
 
 ## Déplacement du film dans la corbeille
-if [[ -e "$actualfile" ]]; then
-    mkdir -p "$trashpath"
-    echo "[INFO] Trashing \"$actualfile\" to \"$trashpath\"" >> "$logfile"
-    mv -f "$actualfile" "$trashpath"
-    if [[ "$?" -ne 0 ]]; then 
-        echo "[ERROR] Abort" >> "$logfile"
-        exit 1
-    fi
-else
-    echo "[INFO] No file to trash \"$actualfile\"" >> "$logfile"
-fi
+trashfiles="$filmspath/$moviename"
+mkdir -p "$trashpath"
+echo "[INFO] Trashing \"$trashfiles\".* to \"$trashpath/\"" >> "$logfile"
+for file in "$trashfiles".*; do
+    echo "[INFO] Trashing \"$file\" to \"$trashpath/\"" >> "$logfile"
+    [[ -e "$file" ]] && mv -f "$file" "$trashpath/" || { echo "[ERROR] Abort" >> "$logfile"; exit 1; }
+done
+
 
 ## Déplacement du film
-destpath="$filmspath/$(basename "$downloadedfile")"
-echo "[INFO] Moving : \"$downloadedfile\" to \"$destpath\"" >> "$logfile"
-mv "$downloadedfile" "$destpath"
-if [[ "$?" -ne 0 ]]; then
-    echo "[ERROR] Abort" >> "$logfile"
-    exit 1
-fi
+[[ "$downloadedfile" == *.* ]] && extension="${downloadedfile##*.}" || extension=""
+destfile="$filmspath/$moviename.$extension"
+echo "[INFO] Moving : \"$downloadedfile\" to \"$destfile\"" >> "$logfile"
+mv -f "$downloadedfile" "$destfile" || { echo "[ERROR] Abort" >> "$logfile"; exit 1; }
+
 
 ## Création du lien radarr (en référentiel docker Radarr)
-echo "[INFO] Creating Radarr link of \"$destpath\" in \"$filmlinkpath\"" >> "$logfile"
+filmlinkpath="$radarrpath/$moviename"
+echo "[INFO] Creating Radarr link of \"$destfile\" in \"$filmlinkpath\"" >> "$logfile"
 mkdir -p "$filmlinkpath"
-ln -s "$destpath" "$filmlinkpath"
-if [[ "$?" -ne 0 ]]; then
-    echo "[ERROR] Continue" >> "$logfile"
-fi
+ln -s "$destfile" "$filmlinkpath" || echo "[ERROR] Continue" >> "$logfile"
+
 
 ## Vider dossier downloads
 dirtemp=$(dirname "$downloadedfile")
 if [[ "$dirtemp" != "$downloadpath" ]]; then
     # Enlever le dossier racine pour obtenir la chaîne des sous-dossiers
-    subdirs=$(echo "$dirtemp" | sed "s|^$downloadpath/||")
     # Extraire le premier sous-dossier (B)
-    first_subdir=$(echo "$subdirs" | cut -d'/' -f1)
+    first_subdir=$(echo "$dirtemp" | sed "s|^$downloadpath/||" | cut -d'/' -f1)
     # Supprimer le dossier et tout son contenu
     echo "[INFO] Deleting \"$first_subdir\" de \"$downloadpath\"" >> "$logfile"
-    rm -r "$downloadpath/$first_subdir"
+    rm -rf "$downloadpath/$first_subdir"
 fi
-
-# ## Vider dossier trash
-# for file in "${deleted_files[@]}"; do
-#     echo "[INFO] Deleting \"$file\"" >> "$logfile"
-#     rm -f "$file"
-# done
 
 ## Arrivé au bout, supprimer le ticket
 echo "[INFO] Remove Ticket \"$ticketfile\"" >> "$logfile"
-rm -f "$ticketfile"
-echo ">>>>>> $(date) End : $filmname" >> "$logfile"
+[[ -f "$ticketfile" ]] && rm -f "$ticketfile"
+echo ">>>>>> $(date) End : $moviename" >> "$logfile"
